# docs/swagger.yaml
basePath: /
definitions:
...
  auth.Role:
    enum:
    - admin
    - member
    - viewer
    - guest
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
    - RoleViewer
    - RoleGuest
...
  interfaces.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/auth.Role'
      updatedAt:
        type: string
    type: object
...
  models.AnalysisReport:
    properties:
      durationMs:
        type: integer
      findings:
        items:
          $ref: '#/definitions/models.Finding'
        type: array
      flowId:
        type: string
      generatedAt:
        type: string
      stats:
        $ref: '#/definitions/models.AnalysisStats'
    type: object
...
  models.AnalysisStats:
    properties:
      blocksAnalyzed:
        type: integer
      errors:
        type: integer
      info:
        type: integer
      rulesRun:
        type: integer
      warnings:
        type: integer
    type: object
...
  models.Finding:
    properties:
      autoFixHint:
        type: string
      blockId:
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      ruleId:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      subflowId:
        type: string
      suggestion:
        type: string
      title:
        type: string
    type: object
...
  models.GraphData:
    properties:
      edges:
        items:
          $ref: '#/definitions/models.GraphEdge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/models.GraphNode'
        type: array
    type: object
  models.GraphEdge:
    properties:
      source:
        type: string
      target:
        type: string
    type: object
  models.GraphNode:
    properties:
      blockCount:
        type: integer
      errorCount:
        type: integer
      id:
        type: string
      label:
        type: string
      type:
        description: '"subflow"'
        type: string
      warnCount:
        type: integer
    type: object
...
  models.RuleConfig:
    properties:
      enabled:
        type: boolean
      options:
        additionalProperties: true
        type: object
      severity:
        type: string
    type: object
...
  models.Severity:
    enum:
    - error
    - warning
    - info
    type: string
    x-enum-varnames:
    - SeverityError
    - SeverityWarning
    - SeverityInfo
...
  models.VariableEvent:
    properties:
      blockId:
        type: string
      line:
        type: integer
      subflowId:
        type: string
      type:
        description: '"init", "mutate", "read"'
        type: string
    type: object
  models.VariableHistory:
    properties:
      events:
        items:
          $ref: '#/definitions/models.VariableEvent'
        type: array
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Pad Analyzer project.
  title: Pad Analyzer API
  version: "1.0"
paths:
  /api/admin/users/{userId}/role:
    post:
      consumes:
      - application/json
      description: Updates the system role for a specific user. Only available to
        system admins.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Set Role Request
        in: body
        name: request
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set user role
      tags:
      - admin
  /api/admin/users/list:
    get:
      description: Returns a list of all registered users. Only available to system
        admins.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/interfaces.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users
      tags:
      - admin
  /api/analysis/analyze:
    post:
      description: Runs all enabled analysis rules on the current flow document.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnalysisReport'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Analyze current flow
      tags:
      - analysis
  /api/analysis/graph:
    get:
      description: Returns a graph representation of the flow's execution path.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GraphData'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get execution graph
      tags:
      - analysis
  /api/analysis/lineage:
    post:
      consumes:
      - application/json
      description: Returns the history and dependencies of a specific variable within
        the flow.
      parameters:
      - description: Variable Lineage Request
        in: body
        name: request
        required: true
        schema:
          properties:
            varName:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VariableHistory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get variable lineage
      tags:
      - analysis
  /api/analysis/rule/config:
    post:
      consumes:
      - application/json
      description: Updates the configuration parameters for a specific analysis rule.
      parameters:
      - description: Update Rule Config Request
        in: body
        name: request
        required: true
        schema:
          properties:
            config:
              $ref: '#/definitions/models.RuleConfig'
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update analysis rule config
      tags:
      - analysis
  /api/analysis/rule/enabled:
    post:
      consumes:
      - application/json
      description: Enables or disables a specific analysis rule by ID.
      parameters:
      - description: Set Rule Enabled Request
        in: body
        name: request
        required: true
        schema:
          properties:
            enabled:
              type: boolean
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Enable/disable analysis rule
      tags:
      - analysis
...
swagger: "2.0"